Revert "Add a bar for instant actions, among other things"
