+++ .gitignore
@@ -1,2 +1,1 @@
 node_modules
-package-lock.json
