+++ chat.html
@@ -3,7 +3,7 @@
   <head>
     <title>My Drive - Google Drive</title>
     <link rel="icon" href="/favicon.ico" type="image/x-icon" />
-    <link rel="stylesheet" href="/raw_chat.css" />
+    <link rel="stylesheet" href="/chat.css" />
   </head>
   <body>
     <!--
@@ -27,6 +27,6 @@
     </div>
     <script src="https://cdn.socket.io/socket.io-1.2.0.js"></script>
     <script src="https://code.jquery.com/jquery-1.11.1.js"></script>
-    <script src="/raw_chat.js"></script>
+    <script src="/chat.js"></script>
   </body>
 </html>
