--- chframe/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>My Drive - Google Drive</title>
-    <link rel="icon" href="/favicon.ico" type="image/x-icon" />
-    <link rel="stylesheet" href="/chat.css" />
-  </head>
-  <body>
-    <div id="topbar"><span>REMOVE&nbsp;TOPBAR</span><a href="https://github.com/nomorenotes/nomorenotes" target=_blank>VIEW&nbsp;CODE</span></div>
-    <iframe id="rawchat" src="/raw_chat">Sorry, your browser does not support IFrames. Click <a href="/raw_chat">here</a> for an iframe-less chat.</iframe>
-    <script src="/chat.js"></script>
-  </body>
-</html>
-
