--- chframe/styles.css
@@ -1,21 +1,0 @@
-div#topbar {
-        height: 100%;
-        width: 5%;
-        left: 0px;
-        top: 0px;
-        position: absolute;
-}
-div#topbar > * {
-        height: 100%;
-        text-align: middle;
-        margin-left: 0.5cm;
-        margin-right: 0.5cm;
-}
-iframe {
-        height: 95%;
-        width: 100%;
-        position: absolute;
-        left: 0px;
-        top: 5%;
-}
-
