--- login/index.html
@@ -1,10 +1,0 @@
-<html>
-  <head>
-    <title>My Drive - Google Drive</title>
-    <link href="/login.css" rel="stylesheet" />
-    <script src="/login.js"></script>
-  </head>
-  <body>
-    <p>This is a stub. Please <a href="javascript:history.go(-1)">go back</a>.<p>
-  </body>
-</html>
