--- login/main.js
@@ -1,1 +1,0 @@
-console.log("Stub page visited");
