--- login/styles.css
@@ -1,6 +1,0 @@
-p {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
