--- prefixes.js
@@ -1,5 +1,0 @@
-module.exports = {
-	"cmdresp": "@",
-	"alert": "!",
-	"msg": "%"
-}
